requirements.txt
coreapi==2.3.3
Django==2.2.6
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
gunicorn==19.9.0
social-auth-app-django==3.1.0
